requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.0
sqlalchemy==2.0.19
pydantic[email]==2.0.3
python-multipart==0.0.6
python-dotenv==1.0.0
pytest==7.4.0
